scripts/csv_to_markdown/requirements.txt
PyYAML==6.0.2
chardet==5.2.0
pandas==2.2.3
tabulate==0.9.0